+++ lib/models/alltrips.dart
@@ -1,0 +1,3 @@
+import './requestdetails.dart';
+
+final List<RequestDetails> allTrips = [];
